requirements.txt
discord.py~=1.6.0
async-property
pyjwt~=2.0.0
sanic~=20.12.1
git+https://github.com/Rapptz/discord-ext-menus
pillow~=8.0.1
pytz~=2020.4
requests
mongoengine
asyncpraw